First Try at API Links
So we can return _links in the JSON (and also show related Resources in the API Explorer). For instance if an endpoint returns a list of accounts _links could show the href and rel=detail for each account. Relations are defined in code in apiRelations. We pass the codeContext (caller, relations and Resource Docs - that contain the URL templates to be used in links) and the dataContext so we can replace placeholder values e.g. BANK_ID with relevant values.