Proposal for connector versionning (example for CreateAccount => AccountRoutings)
We consider in this POC that a connector method is the combination between a method name : ("createAccount") and a version : ("v4.0.0") => connectorMethod (createAccountV400). So, we can give this information (name + version) in method routing system to select the method version we want.
For the JSON Transfer, we consider we can have multiple versions for an OutBound / InBound message. So each connector method can select the version it use for InBound / OutBound messages.